Alotofshit
Caracmenu GUI progressing, now you can press button that do fuckall. it pauses the game for now so we can use it when testing but one day it will only make us unable to move the player and the camera.
stamina work, when sprinting, when chopping, stamina is used. Jump will also use stamina, just didnt do it yet.
Rotation of the wood when picked, you can hold right click and move the mouse if you wanna try, i think it works pretty well.

Working on lvl uping, will probably assign a touch for next update to test the lvl up carac points and all

Also put shaders on the carac menu when opening and closing, for the script in the shader, i just said yes captain and copy pasted